Added extra read and write functionality to check for high out of bounds.

+++ Virtual68HC11/Memory.h
@@ -30,9 +30,9 @@ namespace CPU_6811 {
         
         ~Memory();
         
-        void write(const byte& p_Address, const byte& p_Value);
+        void write(const int& p_Address, const byte& p_Value);
         
-        const byte& read(const byte& p_Address);
+        const byte& read(const int& p_Address);
 
 
 };
